improve XLA custom kernel structure
