+++ brainstate/compile/__init__.py
@@ -17,41 +17,22 @@
 This module contains the functions for the compilation of JAX code.
 """
 
-from ._ad_checkpoint import *
-from ._ad_checkpoint import __all__ as _ad_checkpoint_all
-from ._conditions import *
-from ._conditions import __all__ as _conditions_all
-from ._error_if import *
-from ._error_if import __all__ as _jit_error_all
-from ._jit import *
-from ._jit import __all__ as _jit_all
-from ._loop_collect_return import *
-from ._loop_collect_return import __all__ as _loops_collection
-from ._loop_no_collection import *
-from ._loop_no_collection import __all__ as _loops_no_collection
-from ._make_jaxpr import *
-from ._make_jaxpr import __all__ as _make_jaxpr_all
-from ._progress_bar import *
-from ._progress_bar import __all__ as _progress_bar_all
+from ._ad_checkpoint import checkpoint
+from ._conditions import cond, switch, ifelse
+from ._error_if import jit_error_if
+from ._jit import jit
+from ._loop_collect_return import scan, checkpointed_scan, for_loop, checkpointed_for_loop
+from ._loop_no_collection import while_loop, bounded_while_loop
+from ._make_jaxpr import StatefulFunction, make_jaxpr
+from ._progress_bar import ProgressBar
 
-__all__ = (
-    _jit_error_all
-    + _conditions_all
-    + _make_jaxpr_all
-    + _jit_all
-    + _progress_bar_all
-    + _loops_collection
-    + _loops_no_collection
-    + _ad_checkpoint_all
-)
-
-del (
-    _jit_error_all,
-    _conditions_all,
-    _loops_collection,
-    _make_jaxpr_all,
-    _jit_all,
-    _progress_bar_all,
-    _loops_no_collection,
-    _ad_checkpoint_all
-)
+__all__ = [
+    'checkpoint',
+    'cond', 'switch', 'ifelse',
+    'jit_error_if',
+    'jit',
+    'scan', 'checkpointed_scan', 'for_loop', 'checkpointed_for_loop',
+    'while_loop', 'bounded_while_loop',
+    'StatefulFunction', 'make_jaxpr',
+    'ProgressBar',
+]
